Move instrumentkind_test.go into metric_test.go
